Update files
Added JUnit tests. Delivery schedule is now somewhat functional: it sorts by date, but still needs to be better organized for readability and button actions need to be defined. Updated to launch even if database files are missing. Seems to work fine in each instance except for the StoreSettings.dat being blank. This is creating issues and needs to be debugged.